Moved MADStream implementation to source file
